projects/Merge_images_vertically/requirements.txt
numpy==1.19.1
Pillow==7.2.0